+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v1.7.2
 *In development*
 
 - Move utils.py and path_utils.py files to utils folder maintaining backwards compatibility
+- Fix input text when element has a shadowroot and text contains quotation marks
 
 v1.7.1
 ------